+++ tests/test_endpoints/test_heat_circuit_section.py
@@ -620,7 +620,7 @@ class TestHappyPathHeatCircuitSection:
             (False, HeatCircuitExcessEnergyAvailable.OFF.value, 0),
         ],
     )
-    async def test_get_get_excess_energy_available(
+    async def test_get_excess_energy_available(
         self,
         human_readable: bool,  # noqa: FBT001
         payload_value: str,