+++ tests/test_endpoints/test_hot_water_tank_section.py
@@ -555,7 +555,7 @@ class TestHappyPathHotWaterTankSection:
             (False, HotWaterTankExcessEnergyAvailable.OFF.value, 0),
         ],
     )
-    async def test_get_get_excess_energy_available(
+    async def test_get_excess_energy_available(
         self,
         human_readable: bool,  # noqa: FBT001
         payload_value: str,