+++ CHANGELOG.md
@@ -51,8 +51,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_use_excess_energy()` to the heating circuit
 - Added endpoint `get_use_excess_energy()` to the external heat source
 - Added endpoint `set_use_excess_energy()` to the external heat source
-- Added endpoint `get_min_runtime_excess_energy` to the external heat source
-- Added endpoint `set_min_runtime_excess_energy` to the external heat source
+- Added endpoint `get_min_runtime_excess_energy()` to the external heat source
+- Added endpoint `set_min_runtime_excess_energy()` to the external heat source
+- Added endpoint `get_excess_energy_available()` to the photovoltaics
 
 ## [2.11.0] - 2026-04-13